Hide none schema type
None schema types, like fontVariationSettings, are skipped in rendering
and don't bother with stringifying. This can be used for properties that
are dependent on others (like variable fonts).

In the future we can likely make this fontVariationSettings specific
since this is currently the only place this behavior is implemented.
But this works for now.